re organise constants and utls
