+++ packages/ui-components/package.json
@@ -27,6 +27,7 @@
     "@material-ui/lab": "^4.0.0-alpha.47",
     "@material-ui/pickers": "^3.2.10",
     "@material-ui/styles": "^4.9.10",
+    "@tupaia/utils": "1.0.0",
     "axios": "^0.21.1",
     "date-fns": "^2.12.0",
     "faker": "^4.1.0",
