+++ packages/ui-components/src/components/Map/index.js
@@ -3,9 +3,11 @@
  * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
  *
  */
+export * from './Legend';
 export * from './MapContainer';
-export * from './PolygonLayer';
 export * from './Markers';
 export * from './MarkerLayer';
+export * from './PolygonLayer';
 export * from './TileLayer';
 export * from './TilePicker';
+export * from './utils';
