+++ packages/ui-components/src/components/Map/utils/index.js
@@ -1,0 +1,7 @@
+/*
+ * Tupaia
+ * Copyright (c) 2017 - 2020 Beyond Essential Systems Pty Ltd
+ *
+ */
+export * from './markerColors';
+export * from './markerFormats';
