+++ packages/ui-components/src/components/Map/utils/markerFormats.js
@@ -4,17 +4,17 @@
  *
  */
 
-import { formatDataValue } from '@tupaia/utils';
+import { formatDataValueByType } from '@tupaia/utils';
+import { resolveSpectrumColour } from './markerColors';
 import {
   YES_COLOR,
   NO_COLOR,
   BREWER_AUTO,
   UNKNOWN_COLOR,
-  resolveSpectrumColour,
   MAP_COLORS,
-} from './Markers/markerColors';
-import { SPECTRUM_ICON, DEFAULT_ICON, UNKNOWN_ICON } from './Markers/markerIcons';
-import { SCALE_TYPES } from './constants';
+  SCALE_TYPES,
+} from '../constants';
+import { SPECTRUM_ICON, DEFAULT_ICON, UNKNOWN_ICON } from '../Markers/markerIcons';
 
 export const MEASURE_TYPE_ICON = 'icon';
 export const MEASURE_TYPE_COLOR = 'color';
@@ -109,7 +109,7 @@ function getFormattedValue(value, type, valueInfo, scaleType, valueType, submiss
       if (scaleType === SCALE_TYPES.TIME) {
         return `last submission on ${submissionDate}`;
       }
-      return formatDataValue(value, valueType);
+      return formatDataValueByType({ value }, valueType);
     case MEASURE_TYPE_RADIUS:
     case MEASURE_TYPE_ICON:
     case MEASURE_TYPE_COLOR:
@@ -229,10 +229,9 @@ export function getFormattedInfo(orgUnitData, measureOption) {
     (orgUnitData[displayedValueKey] || orgUnitData[displayedValueKey] === 0)
   ) {
     return {
-      formattedValue: formatDataValue(
-        orgUnitData[displayedValueKey],
+      formattedValue: formatDataValueByType(
+        { value: orgUnitData[displayedValueKey], metadata: orgUnitData.metadata },
         valueType,
-        orgUnitData.metadata,
       ),
       valueInfo,
     };
@@ -256,7 +255,6 @@ export function getFormattedInfo(orgUnitData, measureOption) {
   };
 }
 
-// Todo: refactor hidden measures
 export function getMeasureDisplayInfo(measureData, measureOptions, hiddenMeasures = {}) {
   const displayInfo = {};
 
@@ -350,6 +348,6 @@ export const calculateRadiusScaleFactor = measureData => {
 
 // Take a measureData array where the [key]: value is a number
 // and filters NaN values (e.g. undefined).
-function flattenNumericalMeasureData(measureData, key) {
+export function flattenNumericalMeasureData(measureData, key) {
   return measureData.map(v => parseFloat(v[key])).filter(x => !isNaN(x));
 }
